+++ src/player.c
@@ -1,5 +1,4 @@
 #include "player.h"
-#include "stdio.h"
 
 Player init_player(void)
 {
@@ -7,7 +6,7 @@ Player init_player(void)
     return new_player;
 }
 
-void add_to_hand(Player *player, unsigned int card)
+void add_to_hand(Player *player, unsigned short card)
 {
     player->hand[++player->hand_head] = card;
     score_hand(player);
@@ -16,12 +15,12 @@ void add_to_hand(Player *player, unsigned int card)
 void score_hand(Player *player)
 {
     // count aces last
-    unsigned int ace_count = 0;
+    unsigned short ace_count = 0;
     player->hand_score = 0;
     // add the cards face values to the hand score first
     for (int card = 0; card <= player->hand_head; ++card)
     {
-        unsigned int card_val = (player->hand[card] % 12) + 1;
+        unsigned short card_val = (player->hand[card] % 12) + 1;
         if (card_val == 1)
             ++ace_count;
         else if (card_val > 10)
@@ -30,7 +29,7 @@ void score_hand(Player *player)
             player->hand_score += card_val;
     }
     // then count aces
-    for (unsigned int i = 0; i < ace_count; ++i)
+    for (unsigned short i = 0; i < ace_count; ++i)
     {
         if (player->hand_score <= 10)
             player->hand_score += 11;
