+++ src/player.h
@@ -1,16 +1,21 @@
 #ifndef PLAYER_H
 #define PLAYER_H
 
+#include "stdio.h"
+#include "stdlib.h"
+#include "string.h"
+#include "deck.h"
+
 typedef struct Player
 {
-    unsigned int score;
-    unsigned int hand_score;
+    unsigned short score;
+    unsigned short hand_score;
     int hand_head;
-    unsigned int hand[11];
+    unsigned short hand[11];
 } Player;
 
 Player init_player(void);
-void add_to_hand(Player *player, unsigned int card);
+void add_to_hand(Player *player, unsigned short card);
 void score_hand(Player *player);
 void hand_string_repr(Player *player, char *out);
